Practice 42

+++ Practice41.c
@@ -9,22 +9,22 @@ int main()
     printf("\n");
     printf("=====================================================\n");
 
-    int num1 = 0;
+    float num1 = 0;
     printf("Enter Number 01: ");
-    scanf("%d", &num1);
+    scanf("%f", &num1);
 
-    int num2 = 0;
+    float num2 = 0;
     printf("Enter Number 02: ");
-    scanf("%d", &num2);
+    scanf("%f", &num2);
 
     if (num1 >= num2)
     {
-        printf("\nOf the two, the larger is %d\n", num1);
+        printf("\nOf the two, the larger is %.2f\n", num1);
     }
 
     else
     {
-        printf("\nOf the two, the larger is %d\n", num2);
+        printf("\nOf the two, the larger is %.2f\n", num2);
     }
 
     printf("=====================================================\n");
